Moved testutils to tests dir
