--- tests/common/ping/README.md
@@ -1,2 +1,0 @@
-This function is used to check the host availability in tests.
-Please do not remove.
