--- tests/common/ping/function.json
@@ -1,10 +1,0 @@
-{
-  "scriptFile": "main.py",
-  "bindings": [
-    {
-      "type": "httpTrigger",
-      "direction": "in",
-      "name": "req"
-    }
-  ]
-}
