--- tests/common/ping/main.py
@@ -1,7 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-import azure.functions as func
-
-
-def main(req: func.HttpRequest):
-    return
