+++ tests/endtoend/test_blueprint_functions.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
-from azure_functions_worker import testutils
+from tests.utils import testutils
 
 
 class TestFunctionInBluePrintOnly(testutils.WebHostTestCase):
