+++ tests/endtoend/test_durable_functions.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
-from azure_functions_worker import testutils
+from tests.utils import testutils
 
 
 class TestDurableFunctions(testutils.WebHostTestCase):
