+++ tests/endtoend/test_http_functions.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 import requests
 
-from azure_functions_worker import testutils
+from tests.utils import testutils
 
 REQUEST_TIMEOUT_SEC = 5
