+++ tests/endtoend/test_table_functions.py
@@ -4,7 +4,7 @@ import json
 import pathlib
 import time
 
-from azure_functions_worker import testutils
+from tests.utils import testutils
 
 
 class TestTableFunctions(testutils.WebHostTestCase):
