+++ tests/unittests/azure_namespace_import/azure_namespace_import.py
@@ -5,7 +5,8 @@ import sys
 import shutil
 import asyncio
 
-from azure_functions_worker import protos, testutils
+from azure_functions_worker import protos
+from tests.utils import testutils
 
 
 async def vertify_nested_namespace_import():
