+++ tests/unittests/path_import/path_import.py
@@ -5,7 +5,8 @@ import sys
 import shutil
 import asyncio
 
-from azure_functions_worker import protos, testutils
+from azure_functions_worker import protos
+from tests.utils import testutils
 
 
 async def verify_path_imports():
