+++ tests/unittests/test_file_accessor.py
@@ -6,7 +6,7 @@ import sys
 import unittest
 from unittest import skipIf
 
-from azure_functions_worker import testutils
+from tests.utils import testutils
 from azure_functions_worker.bindings.shared_memory_data_transfer \
     import SharedMemoryException
