+++ tests/unittests/test_mock_blob_shared_memory_functions.py
@@ -11,7 +11,8 @@ from azure_functions_worker.bindings.shared_memory_data_transfer \
     import SharedMemoryMap
 from azure_functions_worker.bindings.shared_memory_data_transfer \
     import SharedMemoryConstants as consts
-from azure_functions_worker import protos, testutils
+from azure_functions_worker import protos
+from tests.utils import testutils
 
 
 @skipIf(sys.platform == 'darwin', 'MacOS M1 machines do not correctly test the'
