+++ tests/unittests/test_mock_log_filtering_functions.py
@@ -2,7 +2,8 @@
 # Licensed under the MIT License.
 from unittest.mock import patch, call
 
-from azure_functions_worker import protos, testutils
+from azure_functions_worker import protos
+from tests.utils import testutils
 from azure_functions_worker.logging import is_system_log_category
