+++ tests/unittests/test_rpc_messages.py
@@ -7,7 +7,8 @@ import tempfile
 import typing
 import unittest
 
-from azure_functions_worker import protos, testutils
+from azure_functions_worker import protos
+from tests.utils import testutils
 from azure_functions_worker.utils.common import is_python_version
