+++ tests/unittests/test_shared_memory_manager.py
@@ -9,7 +9,7 @@ from unittest import skipIf
 from unittest.mock import patch
 from azure_functions_worker.utils.common import is_envvar_true
 from azure.functions import meta as bind_meta
-from azure_functions_worker import testutils
+from tests.utils import testutils
 from azure_functions_worker.bindings.shared_memory_data_transfer \
     import SharedMemoryManager
 from azure_functions_worker.bindings.shared_memory_data_transfer \
