+++ tests/unittests/test_third_party_http_functions.py
@@ -8,8 +8,8 @@ import typing
 import pytest
 from unittest.mock import patch
 
-from azure_functions_worker import testutils
-from azure_functions_worker.testutils import UNIT_TESTS_ROOT
+from tests.utils import testutils
+from tests.utils.testutils import UNIT_TESTS_ROOT
 
 HOST_JSON_TEMPLATE = """\
 {
