+++ tests/unittests/test_utilities_dependency.py
@@ -6,7 +6,7 @@ import importlib.util
 import unittest
 from unittest.mock import patch
 
-from azure_functions_worker import testutils
+from tests.utils import testutils
 from azure_functions_worker.utils.common import is_python_version
 from azure_functions_worker.utils.dependency import DependencyManager
