+++ tests/utils/testutils.py
@@ -49,7 +49,7 @@ from azure_functions_worker.constants import (
 )
 from azure_functions_worker.utils.common import is_envvar_true, get_app_setting
 
-PROJECT_ROOT = pathlib.Path(__file__).parent.parent
+PROJECT_ROOT = pathlib.Path(__file__).parent.parent.parent
 TESTS_ROOT = PROJECT_ROOT / 'tests'
 E2E_TESTS_FOLDER = pathlib.Path('endtoend')
 E2E_TESTS_ROOT = TESTS_ROOT / E2E_TESTS_FOLDER
